Fix default theme's missing updown icon
Remove the files from .gitignore created by genhtml, because these can and
should be created in a specified output directory outside the project
directory.

Remove the godot.info file from .gitignore created by lcov, because this is
an arbitrary filename specified by the user when running lcov.

Add *.gcov files to .gitignore, because these are the files created when
running gcov.